# Conflicts:
#	src/main/java/com/capstone/booking/entity/dto/TicketTypeDTO.java

+++ src/main/java/com/capstone/booking/service/impl/CategoryServiceImpl.java
@@ -41,8 +41,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     //xoa
     @Override
-    public void delete(Long id) {
+    public ResponseEntity<?> delete(long id) {
+        if (!categoryRepository.findById(id).isPresent()) {
+            return new ResponseEntity("Id already exists", HttpStatus.BAD_REQUEST);
+        }
         categoryRepository.deleteById(id);
+        return new ResponseEntity("Delete Successful", HttpStatus.OK);
     }
 
     //thêm
