+++ src/main/java/com/capstone/booking/service/impl/OrderServiceImpl.java
@@ -1,8 +1,8 @@
 package com.capstone.booking.service.impl;
 
+import com.capstone.booking.api.output.Output;
 import com.capstone.booking.common.converter.OrderConverter;
 import com.capstone.booking.common.key.OrderStatus;
-import com.capstone.booking.common.key.PlaceAndGameStatus;
 import com.capstone.booking.entity.Order;
 import com.capstone.booking.entity.User;
 import com.capstone.booking.entity.dto.OrderDTO;
@@ -10,6 +10,7 @@ import com.capstone.booking.repository.OrderRepository;
 import com.capstone.booking.repository.UserRepository;
 import com.capstone.booking.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -52,7 +53,17 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public void delete(long id) {
+    public ResponseEntity<?> delete(long id) {
+        if (!orderRepository.findById(id).isPresent()) {
+            return new ResponseEntity("Id already exists", HttpStatus.BAD_REQUEST);
+        }
         orderRepository.deleteById(id);
+        return new ResponseEntity("Delete Successful", HttpStatus.OK);
+    }
+
+    @Override
+    public ResponseEntity<?> findByStatus(String status, Long limit, Long page) {
+        Output results = orderRepository.findByStatus(status, limit, page);
+        return ResponseEntity.ok(results);
     }
 }
