+++ src/main/java/com/capstone/booking/service/impl/PlaceServiceImpl.java
@@ -13,6 +13,7 @@ import com.capstone.booking.service.PlaceService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -87,8 +88,12 @@ public class PlaceServiceImpl implements PlaceService {
 
     //xóa place
     @Override
-    public void delete(long id) {
+    public ResponseEntity<?> delete(long id) {
+        if (!placeRepository.findById(id).isPresent()) {
+            return new ResponseEntity("Id already exists", HttpStatus.BAD_REQUEST);
+        }
         placeRepository.deleteById(id);
+        return new ResponseEntity("Delete Successful", HttpStatus.OK);
     }
 
     //change status
